+++ apps/utils/image_buffer.h
@@ -135,10 +135,13 @@ OIDN_NAMESPACE_BEGIN
     }
   }
 
+  // Compares an image to a reference image and returns whether they match exactly
+  bool compareImage(const ImageBuffer& image, const ImageBuffer& ref);
+
   // Compares an image to a reference image and returns the number of errors
   // and the average error value
   std::tuple<size_t, double> compareImage(const ImageBuffer& image,
                                           const ImageBuffer& ref,
-                                          double errorThreshold = 0.005);
+                                          double errorThreshold);
 
 OIDN_NAMESPACE_END