+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Version History
 
 ### Changes in v2.3.1:
 
+-   Fixed corrupted output when in-place denoising high-resolution (> 1080p)
+    images where the input and output are stored in different shared buffer
+    objects (created with `oidnNewSharedBuffer*`) that overlap in memory
 -   Fixed issues with cancellation through progress monitor callbacks:
         -   Fixed cancellation requests almost never being fulfilled on CPU
             devices since `v2.3.0`